Modified code to implement modules
