+++ index.html
@@ -33,6 +33,6 @@
         <ul id="user-searches"></ul>
         
       </section>
-      <script src="JS Files/app.js"></script>
+      <script type="module" src="JS Files/app.js"></script>
     </body>
 </html>
